+++ R/wrf_load.R
@@ -227,12 +227,30 @@ wrf_load <- function(
   # Solidify the RasterStack into a more efficient/fast RasterBrick
   rasterBrick <- raster::brick(rasterStack)
   
+  ncAttributes <- nc_getAttributes(nc, print = FALSE)
+  
+  # Generate the RasterBrick title
+  # "PNW-4km_2020090812_07.nc"
+  # "PNW-4km 2020-09-08 12pm, Hour 7"
+  modelNameSubstring <- stringr::str_extract(basename(filePath), "^.*(?=_\\d{10})")
+  modelRunSubstring <- stringr::str_extract(basename(filePath), "\\d{10}(?=_)")
+  modelRunHourSubstring <- stringr::str_extract(basename(filePath), "\\d{2}(?=\\.nc$)")
+  
+  datetime <- lubridate::ymd_h(modelRunSubstring)
+  timeLabel <- strftime(datetime, format = "%Y-%m-%d %I %p", tz = "UTC")
+  
+  title <- paste0(
+    modelNameSubstring, " ",
+    timeLabel, ", forecast hour ", 
+    modelRunHourSubstring
+  )
+  
   # ----- Crop RasterBrick -----------------------------------------------------
   
   if ( is.null(xlim) && is.null(ylim) ) {
     
     # Full grid
-    return(rasterBrick)
+    result <- rasterBrick
     
   } else {
     
@@ -263,8 +281,12 @@ wrf_load <- function(
     ext <- raster::extentFromCells(rasterBrick, cells)
     
     # Crop to xlim, ylim
-    return(raster::crop(rasterBrick, ext))
+    result <- raster::crop(rasterBrick, ext)
     
   }
+  
+  result@title <- title
+  
+  return(result)
   
 }